Fixed some functionaility- and consistency-breaking, plus one mechanics-breaking issue.
fixed the opening of the gui reenabling the program if it was previously disabled, which should not happen. As a rule of thumb, if the program GUI's contents (listviews, buttons etc.) are visible when closing the GUI, the program is active once you close the gui.

* Stopped some notify-calls from triggering when debugging in VSCode.
* added some documentation on specfic lines. More inline-documentation will come

* there are some weird janky recursive loops in some subroutines that could be optimised, but I don't have the time to redesign it and it hasn't broken stuff so far, so that's not a terribly relevant thing. The one part of the program that should be relatively well optimised is as good as it gets with my current understanding of AHK.
* Removed Reload from being a reason to launch the reload script in the restart-functions (f_RestartWithLastBundle, fRestartEmpty), because that is... pointless? Just confuses people if stuff restarts _twice_.

* replaced restart-scripts with restart-exes. Because this program is intended to be run in isolation on computers without ahk installs as well,I see no reason to launch the ahk-versions. At the same time, the original scripts of those executables wil remain as source-files in the project, for nothing else but to give easier ways of inspecting their contents.
---
* I hope the LogError function doesn't spam error-message-boxes anymore. I generally hope the program just restarts on errors.
---
Known bugs:
the function for fetching URL's can sometimes freeze up, at least on my machine.

If the program suddenly starts websites that are _clearly_ not ment to be closed, enable Debug-mode and check if the "Current URL" remains the same regardless on which actually active tab/URL you test it on. In this case, restart the program. This hasn't happened to me too often, and I have no idea why it happens so far. Currently time is sparse, so the current tip is to simply restart the script and hope for the best. If anyone who is more informed on the workings of the GetURL- function and why this error happens, get in contact please.